requirements.txt
Django
djangorestframework
drf-spectacular
psycopg2-binary
celery
redis
requests
httpx
langchain
langchain_openai